muitas alteraçoes acumuladas, mas agora implementando JWT e argon2 funcioanndo

+++ DTO/RelatorioDto.cs
@@ -2,13 +2,13 @@
 {
     public class RelatorioDto
     {
-        public long TerrenoId { get; set; }
+        public int TerrenoId { get; set; }
         public string ImagemBase64 { get; set; }
         public string Descricao { get; set; }
 
         public RelatorioDto() { }
 
-        public RelatorioDto(long terrenoId, string imagemBase64, string descricao)
+        public RelatorioDto(int terrenoId, string imagemBase64, string descricao)
         {
             TerrenoId = terrenoId;
             ImagemBase64 = imagemBase64;
